requirements.txt
# WawaTrader - LLM-Powered Trading System
# Core Dependencies

# Trading & Market Data
alpaca-trade-api>=3.0.0
alpaca-py>=0.9.0

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# LLM Integration (for LM Studio via OpenAI-compatible API)
openai>=1.0.0

# Configuration Management
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# Utilities
requests>=2.31.0
aiohttp>=3.8.0
python-dateutil>=2.8.0
pytz>=2023.3

# Logging & Monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
